requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Mongo driver
pymongo

# Environment file handling
python-dotenv

# REST Calls
requests

# OpenAI
OpenAI

# Mistral AI
mistralai

